Created a new model for articles to be used in the application. This model will help in managing the data structure and ensure consistency.
