+++ frontend/scribere/src/app/pages/article/article.component.ts
@@ -1,6 +1,7 @@
 import { Component, inject, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { ArticleService, Article } from '../../services/article.service';
+import { ArticleService } from '../../services/article.service';
+import { Article } from '../../model/article.model';
 import { CommonModule } from '@angular/common';
 import { finalize } from 'rxjs';
